Add structure of State library with unit tests
Unit tests are performed using Minitest.

Notice that the file State.cpp needs to be included in StateTest.cpp
because the header file does not define the State class but only
declares it.

I am not quit sure why the same doesn't cause problems
when building Arduino code, but I suspect that the default Arduino
compiler (linker) options may have something to do with that, because
a workaround to including the file in our case would be to add
State.cpp as an argument when compiling StateTest.cpp.

I prefer the explicit include for now.