Refactor group tests by input rather that initial state
Features (e.g. new buttons) will likely be implemented progressively
as new factors will be allowed to trigger state transitions.
(An example could be the addition of a new button.)

Beside, since people provide input, I believe the test suite will be
easier to understand if the entry point is the input, and not the rather
abstract collection of states.